backend/requirements.txt
python-dotenv>=1.0.0
langgraph>=0.0.10
mcp>=0.2.0
langchain-mcp-adapters>=0.0.1
langchain-openai>=0.0.5 